Update test structure and remove unwanted dependencies
