+++ docs/conf.py
@@ -26,7 +26,7 @@ autoapi_dirs = ["../src"]
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", ".pytest_cache"]
 
 # -- Options for HTML output -------------------------------------------------
